requirements.txt
addict==2.4.0
click==8.1.3
h5py==3.8.0
numpy==1.26.1
nystrom-attention==0.0.11
opencv-python==4.8.0.74
openslide-python==1.2.0
pandas==2.0.0
Pillow==9.5.0
pytorch-lightning==2.0.8
PyYAML==6.0
scikit-image==0.21.0
timm==1.0.8
torch==1.13.0
torchmetrics==1.1.2
torchvision==0.14.0
wandb==0.18.7
